move `draft mode` functionality into mixin class Since a few nodes are use these methods no sense of mixing them with others in base class.
