+++ nodes/curve/length_parameter.py
@@ -7,8 +7,10 @@ from bpy.props import FloatProperty, EnumProperty, BoolProperty, IntProperty
 from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, zip_long_repeat, ensure_nesting_level
 from sverchok.utils.curve import SvCurveLengthSolver, SvCurve
+from sverchok.utils.nodes_mixins.draft_mode import DraftMode
 
-class SvCurveLengthParameterNode(bpy.types.Node, SverchCustomTreeNode):
+
+class SvCurveLengthParameterNode(DraftMode, bpy.types.Node, SverchCustomTreeNode):
     """
     Triggers: Curve Length Parameter
     Tooltip: Solve curve length (natural) parameter
